[chore #159102302] : Rename API entry point
