--- API/.gitignore
@@ -1,8 +1,0 @@
-./
-apiEnv/
-.vscode/
-app/__pycache__/
-*.pyc
-.pytest_cache/
-.cache/
-
