+++ requirements.txt
@@ -20,3 +20,29 @@ six==1.11.0
 typed-ast==1.1.0
 Werkzeug==0.14.1
 wrapt==1.10.11
+astroid==2.0
+atomicwrites==1.1.5
+attrs==18.1.0
+autopep8==1.3.5
+click==6.7
+coverage==4.5.1
+Flask==1.0.2
+gunicorn==19.9.0
+isort==4.3.4
+itsdangerous==0.24
+Jinja2==2.10
+lazy-object-proxy==1.3.1
+MarkupSafe==1.0
+mccabe==0.6.1
+more-itertools==4.2.0
+pkg-resources==0.0.0
+pluggy==0.6.0
+py==1.5.4
+pycodestyle==2.4.0
+pylint==2.0.0
+pytest==3.6.3
+rope==0.10.7
+six==1.11.0
+typed-ast==1.1.0
+Werkzeug==0.14.1
+wrapt==1.10.11
